CDAP-9283 Background upgrade thread

+++ cdap-app-fabric/src/main/java/co/cask/cdap/internal/app/runtime/schedule/store/DatasetBasedStreamSizeScheduleStore.java
@@ -25,6 +25,7 @@ import co.cask.cdap.api.dataset.table.Table;
 import co.cask.cdap.api.schedule.SchedulableProgramType;
 import co.cask.cdap.common.logging.LogSamplers;
 import co.cask.cdap.common.logging.Loggers;
+import co.cask.cdap.common.utils.ProjectInfo;
 import co.cask.cdap.internal.app.runtime.schedule.AbstractSchedulerService;
 import co.cask.cdap.internal.app.runtime.schedule.StreamSizeScheduleState;
 import co.cask.cdap.internal.schedule.StreamSizeSchedule;
@@ -32,8 +33,9 @@ import co.cask.cdap.proto.ProgramType;
 import co.cask.cdap.proto.id.ApplicationId;
 import co.cask.cdap.proto.id.ProgramId;
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.base.Joiner;
 import com.google.common.base.Preconditions;
+import com.google.common.cache.CacheBuilder;
+import com.google.common.cache.LoadingCache;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
@@ -42,18 +44,21 @@ import com.google.gson.reflect.TypeToken;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
 import org.apache.tephra.TransactionAware;
+import org.apache.tephra.TransactionConflictException;
 import org.apache.tephra.TransactionExecutor;
 import org.apache.tephra.TransactionExecutorFactory;
 import org.apache.tephra.TransactionFailureException;
+import org.apache.tephra.TransactionNotInProgressException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.lang.reflect.Type;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicInteger;
 import javax.annotation.Nullable;
 
 /**
@@ -62,6 +67,7 @@ import javax.annotation.Nullable;
 @Singleton
 public class DatasetBasedStreamSizeScheduleStore {
   public static final String KEY_PREFIX = "streamSizeSchedule";
+  private static final String NAME = "StreamSizeScheduleStore";
 
   private static final Logger LOG = LoggerFactory.getLogger(DatasetBasedStreamSizeScheduleStore.class);
   // Limit the number of logs when the rows have invalid columns. This will be called only during the startup when the
@@ -70,6 +76,8 @@ public class DatasetBasedStreamSizeScheduleStore {
   private static final Logger LIMITED_LOG = Loggers.sampling(LOG, LogSamplers.onceEvery(100));
 
   private static final Gson GSON = new Gson();
+  private static final byte[] APP_VERSION_UPGRADE_KEY = Bytes.toBytes("version.streamsize.schedule");
+  private static final byte[] COLUMN = Bytes.toBytes('c');
   private static final byte[] SCHEDULE_COL = Bytes.toBytes("schedule");
   private static final byte[] BASE_SIZE_COL = Bytes.toBytes("baseSize");
   private static final byte[] BASE_TS_COL = Bytes.toBytes("baseTs");
@@ -81,12 +89,16 @@ public class DatasetBasedStreamSizeScheduleStore {
 
   private final TransactionExecutorFactory factory;
   private final ScheduleStoreTableUtil tableUtil;
+  private final AtomicBoolean storeInitialized;
+
+  private LoadingCache<byte[], Boolean> upgradeCacheLoader;
   private Table table;
 
   @Inject
   public DatasetBasedStreamSizeScheduleStore(TransactionExecutorFactory factory, ScheduleStoreTableUtil tableUtil) {
     this.tableUtil = tableUtil;
     this.factory = factory;
+    this.storeInitialized = new AtomicBoolean(false);
   }
 
   /**
@@ -96,6 +108,10 @@ public class DatasetBasedStreamSizeScheduleStore {
     table = tableUtil.getMetaTable();
     Preconditions.checkNotNull(table, "Could not get dataset client for data set: %s",
                                ScheduleStoreTableUtil.SCHEDULE_STORE_DATASET_NAME);
+    upgradeCacheLoader = CacheBuilder.newBuilder()
+      .expireAfterWrite(1, TimeUnit.MINUTES)
+      .build(new UpgradeValueLoader(NAME, factory, table, storeInitialized));
+    storeInitialized.set(true);
   }
 
   /**
@@ -222,15 +238,17 @@ public class DatasetBasedStreamSizeScheduleStore {
    */
   public synchronized void delete(final ProgramId programId, final SchedulableProgramType programType,
                                   final String scheduleName) throws InterruptedException, TransactionFailureException {
+    final boolean needVersionLessDelete = !isUpgradeComplete();
     factory.createExecutor(ImmutableList.of((TransactionAware) table))
       .execute(new TransactionExecutor.Subroutine() {
         @Override
         public void apply() throws Exception {
           String rowKey = getRowKey(programId, programType, scheduleName);
-
-          String versionLessRowKey = removeAppVersion(rowKey);
-          if (versionLessRowKey != null) {
-            table.delete(Bytes.toBytes(versionLessRowKey));
+          if (needVersionLessDelete) {
+            String versionLessRowKey = removeAppVersion(rowKey);
+            if (versionLessRowKey != null) {
+              table.delete(Bytes.toBytes(versionLessRowKey));
+            }
           }
           table.delete(Bytes.toBytes(rowKey));
         }
@@ -363,25 +381,80 @@ public class DatasetBasedStreamSizeScheduleStore {
   /**
    * Method to add version in StreamSizeSchedule row key in SchedulerStore.
    *
-   * @throws Exception
+   * @throws InterruptedException
+   * @throws IOException
+   * @throws DatasetManagementException
    */
-  public void upgrade()
-    throws InterruptedException, TransactionFailureException, IOException, DatasetManagementException {
-    initialize();
-    factory.createExecutor(ImmutableList.of((TransactionAware) table))
-      .execute(new TransactionExecutor.Subroutine() {
-        @Override
-        public void apply() {
-          upgradeVersionKeys();
+  public void upgrade() throws InterruptedException, IOException, DatasetManagementException {
+    // Wait until the store is initialized
+    while (!storeInitialized.get()) {
+      TimeUnit.SECONDS.sleep(10);
+    }
+
+    if (isUpgradeComplete()) {
+      LOG.info("{} is already upgraded.", NAME);
+      return;
+    }
+
+    final AtomicInteger maxNumberUpdateRows = new AtomicInteger(1000);
+    final AtomicInteger sleepTimeInSecs = new AtomicInteger(60);
+    LOG.info("Starting upgrade of {}.", NAME);
+    while (!isUpgradeComplete()) {
+      sleepTimeInSecs.set(60);
+      try {
+        factory.createExecutor(ImmutableList.of((TransactionAware) table))
+          .execute(new TransactionExecutor.Subroutine() {
+            @Override
+            public void apply() {
+              if (upgradeVersionKeys(table, maxNumberUpdateRows.get())) {
+                // Upgrade is complete. Mark that app version upgrade is complete in the table.
+                table.put(APP_VERSION_UPGRADE_KEY, COLUMN, Bytes.toBytes(ProjectInfo.getVersion().toString()));
+              }
+            }
+          });
+      } catch (TransactionFailureException e) {
+        if (e instanceof TransactionConflictException) {
+          LOG.debug("Upgrade step faced Transaction Conflict exception. Retrying operation now.", e);
+          sleepTimeInSecs.set(10);
+        } else if (e instanceof TransactionNotInProgressException) {
+          int currMaxRows = maxNumberUpdateRows.get();
+          if (currMaxRows > 500) {
+            maxNumberUpdateRows.decrementAndGet();
+          } else {
+            LOG.warn("Could not complete upgrade of {}, tried for 500 times", NAME);
+            return;
+          }
+          sleepTimeInSecs.set(10);
+          LOG.debug("Upgrade step faced a Transaction Timeout exception. " +
+                      "Current number of max update rows is set to : {} and retrying the operation now.",
+                    maxNumberUpdateRows.get(), e);
+        } else {
+          LOG.error("Upgrade step faced exception. Will retry operation after some delay.", e);
+          sleepTimeInSecs.set(60);
         }
-      });
+      }
+      TimeUnit.SECONDS.sleep(sleepTimeInSecs.get());
+    }
+    LOG.info("Upgrade of {} is complete.", NAME);
   }
 
-  private void upgradeVersionKeys() {
-    Joiner joiner = Joiner.on(":");
+  // Returns true if the upgrade flag is set. Upgrade could have completed earlier than this since this flag is
+  // updated asynchronously.
+  public boolean isUpgradeComplete() {
+    if (upgradeCacheLoader == null) {
+      return false;
+    }
+    return upgradeCacheLoader.getUnchecked(APP_VERSION_UPGRADE_KEY);
+  }
+
+  // Return whether the upgrade process is complete - determined by checking if there were no rows that were
+  // upgraded after the invocation of this method.
+  private boolean upgradeVersionKeys(Table table, int maxNumberUpdateRows) {
+    int numRowsUpgraded = 0;
     try (Scanner scan = getScannerWithPrefix(KEY_PREFIX)) {
       Row next;
-      while ((next = scan.next()) != null) {
+      // Upgrade only N rows in one transaction to reduce the probability of conflicts with regular Store operations.
+      while (((next = scan.next()) != null) && (numRowsUpgraded < maxNumberUpdateRows)) {
         if (isInvalidRow(next)) {
           LIMITED_LOG.debug("Stream Sized Schedule entry with Row key {} does not have all columns.",
                             Bytes.toString(next.getRow()));
@@ -393,22 +466,32 @@ public class DatasetBasedStreamSizeScheduleStore {
         // Row key for the trigger should be of the form -
         // streamSizeSchedule:namespace:application:type:program:schedule
         if (splits.length != 6) {
-          LOG.debug("Skip upgrading StreamSizeSchedule {}. Expected row key " +
+          LIMITED_LOG.debug("Skip upgrading StreamSizeSchedule {}. Expected row key " +
                      "format 'streamSizeSchedule:namespace:application:type:program:schedule'", oldRowKeyString);
           continue;
         }
-        List<String> splitsList = new ArrayList<>(Arrays.asList(splits));
         // append application version after application name
-        splitsList.add(3, ApplicationId.DEFAULT_VERSION);
-        String newRowKeyString = joiner.join(splitsList);
-        byte[] newRowKey = Bytes.toBytes(newRowKeyString);
+        byte[] newRowKey = Bytes.toBytes(ScheduleUpgradeUtil.getNameWithDefaultVersion(splits, 3));
+
+        // Check if a newRowKey is already present, if it is present, then simply delete the oldRowKey and continue;
+        Row row = table.get(newRowKey);
+        if (!row.isEmpty()) {
+          table.delete(oldRowKey);
+          numRowsUpgraded++;
+          continue;
+        }
+
         Put put = new Put(newRowKey);
         for (Map.Entry<byte[], byte[]> colValEntry : next.getColumns().entrySet()) {
           put.add(colValEntry.getKey(), colValEntry.getValue());
         }
         table.put(put);
         table.delete(oldRowKey);
+        numRowsUpgraded++;
       }
     }
+
+    // If no rows were upgraded, notify that the upgrade process has completed.
+    return (numRowsUpgraded == 0);
   }
 }
